cleaned app.js
