+++ package.json
@@ -9,6 +9,7 @@
     "mapbox-gl": "^1.13.0",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.1",
     "web-vitals": "^0.2.4"
   },
