--- src/components/LocationList.js
@@ -1,12 +1,0 @@
-import { useState } from 'react'
-
-export default function LocationList () {
-  const [locations, setLocations] useState([
-    {-78.8986, 35.9940},
-    {-78.8990, 35.9950},
-    {-78.8980, 35.9930}
-  ])
-  return (
-    
-  )
-}
